+++ lib/karafka/active_job/routing/extensions.rb
@@ -10,9 +10,14 @@ module Karafka
         # This method simplifies routes definition for ActiveJob topics / queues by auto-injecting
         # the consumer class
         # @param name [String, Symbol] name of the topic where ActiveJobs jobs should go
-        def active_job_topic(name)
+        # @param block [Proc] block that we can use for some extra configuration
+        def active_job_topic(name, &block)
           topic(name) do
             consumer App.config.internal.active_job.consumer
+
+            next unless block
+
+            instance_eval(&block)
           end
         end
       end