+++ lib/karafka/connection/client.rb
@@ -86,8 +86,7 @@ module Karafka
       # @param message [Karafka::Messages::Message]
       def store_offset(message)
         @mutex.synchronize do
-          @offsetting = true
-          @kafka.store_offset(message)
+          internal_store_offset(message)
         end
       end
 
@@ -104,14 +103,7 @@ module Karafka
       def commit_offsets(async: true)
         @mutex.lock
 
-        return unless @offsetting
-
-        @kafka.commit(nil, async)
-        @offsetting = false
-      rescue Rdkafka::RdkafkaError => e
-        return if e.code == :no_offset
-
-        raise e
+        internal_commit_offsets(async: async)
       ensure
         @mutex.unlock
       end
@@ -128,7 +120,11 @@ module Karafka
       #
       # @param message [Messages::Message, Messages::Seek] message to which we want to seek to
       def seek(message)
+        @mutex.lock
+
         @kafka.seek(message)
+      ensure
+        @mutex.unlock
       end
 
       # Pauses given partition and moves back to last successful offset processed.
@@ -144,15 +140,19 @@ module Karafka
         # Do not pause if the client got closed, would not change anything
         return if @closed
 
+        pause_msg = Messages::Seek.new(topic, partition, offset)
+
+        # Since we want to pause on a given offset, we commit it
+        internal_store_offset(pause_msg)
+        internal_commit_offsets(async: false)
+
         tpl = topic_partition_list(topic, partition)
 
         return unless tpl
 
         @kafka.pause(tpl)
 
-        pause_msg = Messages::Seek.new(topic, partition, offset)
-
-        seek(pause_msg)
+        @kafka.seek(pause_msg)
       ensure
         @mutex.unlock
       end
@@ -165,6 +165,11 @@ module Karafka
         @mutex.lock
 
         return if @closed
+
+        # Always commit synchronously offsets if any when we resume
+        # This prevents resuming without offset in case it would not be committed prior
+        # We can skip performance penalty since resuming should not happen too often
+        internal_commit_offsets(async: false)
 
         tpl = topic_partition_list(topic, partition)
 
@@ -214,11 +219,31 @@ module Karafka
 
       private
 
+      # Non thread-safe offset storing method
+      # @param message [Karafka::Messages::Message]
+      def internal_store_offset(message)
+        @offsetting = true
+        @kafka.store_offset(message)
+      end
+
+      # Non thread-safe message committing method
+      # @param async [Boolean] should the commit happen async or sync (async by default)
+      def internal_commit_offsets(async: true)
+        return unless @offsetting
+
+        @kafka.commit(nil, async)
+        @offsetting = false
+      rescue Rdkafka::RdkafkaError => e
+        return if e.code == :no_offset
+
+        raise e
+      end
+
       # Commits the stored offsets in a sync way and closes the consumer.
       def close
-        commit_offsets!
-
         @mutex.synchronize do
+          internal_commit_offsets(async: false)
+
           @closed = true
 
           # Remove callbacks runners that were registered