+++ lib/karafka/pro/processing/jobs/consume_non_blocking.rb
@@ -1,21 +1,20 @@
 # frozen_string_literal: true
 
+# This Karafka component is a Pro component.
+# All of the commercial components are present in the lib/karafka/pro directory of this
+# repository and their usage requires commercial license agreement.
+#
+# Karafka has also commercial-friendly license, commercial support and commercial components.
+#
+# By sending a pull request to the pro components, you are agreeing to transfer the copyright of
+# your code to Maciej Mensfeld.
+
 module Karafka
   module Pro
     # Pro components related to processing part of Karafka
     module Processing
       # Pro jobs
       module Jobs
-        # This Karafka component is a Pro component.
-        # All of the commercial components are present in the lib/karafka/pro directory of this
-        # repository and their usage requires commercial license agreement.
-        #
-        # Karafka has also commercial-friendly license, commercial support and commercial
-        # components.
-        #
-        # By sending a pull request to the pro components, you are agreeing to transfer the
-        # copyright of your code to Maciej Mensfeld.
-
         # The main job type in a non-blocking variant.
         # This variant works "like" the regular consumption but pauses the partition for as long
         # as it is needed until a job is done.