+++ spec/integrations/pausing/seeking_offset_after_pausing.rb
@@ -6,8 +6,8 @@
 setup_karafka do |config|
   config.max_messages = 5
   config.pause_timeout = 5_000
-  config.pause_max_timeout = 10_000
-  config.pause_with_exponential_backoff = true
+  config.pause_max_timeout = 25_000
+  config.pause_with_exponential_backoff = false
   config.manual_offset_management = true
 end
 
@@ -17,9 +17,7 @@ class Consumer < Karafka::BaseConsumer
 
     unless @paused
       @paused = true
-      # Skip one from the next batch
       pause(messages.last.offset + 5)
-      sleep 2
       seek(@first_message.offset)
     end
 
@@ -34,8 +32,8 @@ draw_routes(Consumer)
 20.times { |i| produce(DataCollector.topic, i.to_s) }
 
 start_karafka_and_wait_until do
-  DataCollector.data[:messages].size >= 25
+  DataCollector.data[:messages].size > 20
 end
 
-assert_equal 25, DataCollector.data[:messages].size
+assert_equal true, DataCollector.data[:messages].size > 20
 assert_equal 20, DataCollector.data[:messages].uniq.size